[refactor] #133 SwiftSoup Live, Test 분리
