+++ Projects/CoreKit/Sources/Data/Client/SwiftSoupClient/SwiftSoupClient+TestKey.swift
@@ -1,0 +1,15 @@
+//
+//  SwiftSoupClient+TestKey.swift
+//  CoreKit
+//
+//  Created by 김민호 on 9/30/24.
+//
+
+import Foundation
+
+import Dependencies
+
+extension SwiftSoupClient: TestDependencyKey {
+    public static let testValue = Self()
+    public static let previewValue = liveValue
+}
